marchandise and categorie done

+++ src/main/java/com/izorai/pfa/module2/DTO/marchandises/CategorieDTO.java
@@ -1,4 +1,7 @@
 package com.izorai.pfa.module2.DTO.marchandises;
 
-public record CategorieDTO(int id, String categorie, String description) {}
+public record CategorieDTO(
+        Long id,
+        String categorie,
+        String description) {}
