+++ src/main/java/com/izorai/pfa/module2/DTO/marchandises/UniteDTO.java
@@ -1,4 +1,4 @@
 package com.izorai.pfa.module2.DTO.marchandises;
 
-public record UniteDTO(int id, String unite, String description) {}
+public record UniteDTO(Long id, String unite, String description) {}
